feat(Pointer): separate play area cursor into own script
This commit brings some breaking changes to the World Pointer script.

The Play Area Cursor used to be part of the World Pointer script but
this meant the play area cursor was coupled to the pointer script and
meant that adding custon play area cursors would mean changing core
code.

A new `VRTK_PlayAreaCursor` script has now been added that can be
applied to the same game object that a World Poiinter component is
applied to. If the Play Area Cursor is applied then it works the same
way as previously, with the same options just moved into the new
script.
