+++ Assets/VRTK/Examples/Resources/Scripts/Setup_Scene_031.cs
@@ -12,12 +12,15 @@
             if (!initalised)
             {
                 var headset = VRTK_DeviceFinder.HeadsetTransform();
+                if (!headset.GetComponent<VRTK_PlayAreaCursor>())
+                {
+                    headset.gameObject.AddComponent<VRTK_PlayAreaCursor>();
+                }
                 if (!headset.GetComponent<VRTK_BezierPointer>())
                 {
                     var pointer = headset.gameObject.AddComponent<VRTK_BezierPointer>();
 
                     pointer.controller = VRTK_DeviceFinder.GetControllerRightHand().GetComponent<VRTK_ControllerEvents>();
-                    pointer.showPlayAreaCursor = true;
                     pointer.pointerVisibility = VRTK_WorldPointer.pointerVisibilityStates.Always_Off;
                     pointer.pointerLength = 7f;
                     pointer.pointerDensity = 1;
