+++ Assets/VRTK/Scripts/VRTK_PlayAreaCursor.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 9d7236ea6e01c074ba249e564af997a0
+timeCreated: 1475140565
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
